+++ csharp/ToolGood.FlowVision/ToolGood.Algorithm2/LitJson/JsonMapper.cs
@@ -18,7 +18,7 @@ namespace ToolGood.Algorithm2.LitJson
 			}
 
 			if (reader.Token == JsonToken.Double) {
-				instance.SetDouble((double)reader.Value);
+				instance.SetDouble((decimal)reader.Value);
 				return instance;
 			}
